@@ -1,6 +1,6 @@
 package net.guides.springboot.loginregistrationspringbootauthjsp.model;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.Set;
 
 @Entity
