@@ -1,12 +1,9 @@
 package net.guides.springboot2.springboot2webappjsp;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest
+@SpringBootTest(classes = TestSpringboot2WebappJspApplication.class)
 public class Springboot2WebappJspApplicationTests {
 
 	@Test
