@@ -3,7 +3,7 @@ package net.javaguides.springboot;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@SpringBootTest(classes = TestSpringbootHibernateOneManyMappingApplication.class)
 class SpringbootHibernateOneManyMappingApplicationTests {
 
 	@Test
