@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 public class HelloController {
 
 	@GetMapping({"/", "/hello"})
-	public String hello(@RequestParam(value = "name", defaultValue = "World",
-	required = true) String name, Model model) {
+	public String hello(@RequestParam(defaultValue = "World",
+            required = true) String name, Model model) {
 		model.addAttribute("name", name);
 		return "hello";
 	}
